pyproject.toml
[project]
name = "hmn-2025-spring-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.14",
    "numpydantic>=1.6.8",
    "openai>=1.64.0",
    "opencv-contrib-python>=4.11.0.86",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pytube3>=9.6.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "youtube-transcript-api>=0.6.3",
]
